requirements.txt
fastapi
uvicorn[standard]
gunicorn
httpx
python-dotenv
pydantic
beautifulsoup4
python-docx
PyMuPDF
langchain
langchain-community
langchain-google-genai
sentence-transformers
faiss-cpu